script: Make list-issues-for-pr reliable
Fix the `list-issues-for-pr` command to correctly return the list of
issues for the specified PR number. The GitHub API appears to be either
broken, limited or badly documented. Hence, we are forced to manually
extract the issue numbers from the PR.

Also marked the `list-issue-linked-prs` and `list-pr-linked-issues`
commands as unreliable, again due to seeming GitHub API limitations.

Fixes: #16.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>