scripts: Don't require leading hash for fixes issue number
The "Fixes" comment in a PR is supposed to specify an issue number with
a leading `#` (such as `Fixes: #123`). However, we have already merged
PR without the `#`. This breaks the hub util scripts
`list-issues-for-pr` command, so update that script to accommodate this
awkward PRs and allow the related issues to be displayed.

Fixes: #28.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>